Hallway And Shark Hit
I added the ability for the shark to hit down doors, it just requires the shark to have the shark tag. I also added the objects to the hallway. By noon, I plan to allow the hallway objects to be destroyed by the shark and add the buttons to create flares to distract the shark in the final room.